A low-hanging fruit among many (scheduling) ideas!

---

For big state transitions:

<img width="1269" height="190" alt="Screenshot From 2025-09-26 16-03-34" src="https://github.com/user-attachments/assets/c933c627-838b-4d8c-8e71-a2d2feb1cd95" />

---

For small state transitions:

`reth-bench` for the first 300 blocks of Shanghai:

```
$ reth-bench new-payload-fcu --advance 300
```

aa192c255b53335ba509987c82ab26e79e954195 (`main`):
```
2025-09-26T09:06:00.711782Z  INFO Total Ggas/s: 0.7019 total_duration=6.539663621s total_gas_used=4590139398 blocks_processed=300
```

This branch on aa192c255b53335ba509987c82ab26e79e954195:
```
2025-09-26T09:23:34.656611Z  INFO Total Ggas/s: 0.7467 total_duration=6.146954381s total_gas_used=4590139398 blocks_processed=300
```

Speedup:
- Ggas/s: 0.7019 -> 0.7467 ~ 6% higher execution throughput
- Duration: 6.539663621s -> 6.146954381s ~ 6% lower payload latency
- Best win: From 32.3ms down to 22.4ms payload latency